# server/docs/swagger.yaml
basePath: /api
host: 127.0.0.1:8080
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Aldb Example API
  version: "1.0"
paths:
...
  /alga/anno:
    get:
      consumes:
      - application/json
      description: 藻类图像获取标注
      parameters:
      - description: 藻类名称
        in: query
        name: alga
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{code, msg, data}'
          schema:
            type: object
      summary: GetAnnotationByAlga
      tags:
      - aldb
...
  /user/anno:
    get:
      consumes:
      - application/json
      description: 用户获取个人标注
      parameters:
      - description: 用户邮箱
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{code, msg, data}'
          schema:
            type: object
      summary: GetAnnotationByUser
      tags:
      - aldb
...
swagger: "2.0"